app/requirements.txt
psycopg2-binary
pymongo
redis
pandas